WIP: updates to latest dependencies
This attempts to update to latest dependencies. There is a glitch
locally with one test.

Also, controlplane's `istio_test` `Test_IstioGateway` is untranslatable
into current apis. We need to decide whether to create fixtures to
re-implement it, or leave it out. Considering it was already with a
hazard sign, I figured worthwhile asking:

```
// NOTE: This test will fail on macOS due to an issue with Envoy, the same issue as debug logging
```

Signed-off-by: Adrian Cole <adrian@tetrate.io>