Migrates e2e tests from ginkgo,gomega to testing,testify (#130)
To call ginkgo,gomega indirect would be putting things lightly. A combination of an indirect testing library and long-running tests are a recipe for frustration. I've begun a process to remove ginkgo,gomega with a couple tests
that I had been looking at.

See #127

## Some benefits
* Before, we both had tests using normal testing,testify and also ginkgo,gomega. Now we only have way to write tests: testing,testify.
* testing,testify are more popular and easier to learn than ginkgo,gomega. testing is the standard library described in all golang books.  testify is a simple assertion library 10x more popular than gomega in terms of star count. Neither require learning BDD concepts.
* Before, our running of e2e tests was significantly different than our other non-gingko tests in the same project. It literally required compilation of a platform-specific binary named 'e2e', adding a lot of complexity to CI and local setup.
* Before, individual tests cannot be executed via normal means, such as Goland. Instead, you have to run the suite file. Now, you can execute and more importantly debug individual tests instead of via a suite wrapper.
* Before, writing tests not only required understanding of the subject, but also the ginkgo runtime. Otherwise, statements like `var _ = Describe("getenvoy extension init", func() {` seem pure magic.
* Before, the nature of test execution required a combination of understanding the go platform and also ginkgo
* Before, updating Ginkgo/gomega would skew the dependency tree, due to their dependencies on two protobuf libraries. Now, the test runtime does not interfere with main code.
* Before, the same test took more lines of code to achieve the same table test. Less code is less to maintain.
* Some other notes about migration off are captured in this issue https://github.com/openservicemesh/osm/issues/1704

## Notable differences

Gomega had a different way of logging. For example instead of using a logging library, it would use statements like this:
`By("changing to the output directory")` which translate into output `STEP: changing to the output directory`. This is mainly to satisfy BDD idioms and is meant to be interpreted amongst the many lines of normal log output. For cases where the points inside the test should be logged, we should use normal logging instead. However, I don't recommend making log statements unless they are meaningful and clarify progress.

Also, those not used to it will notice use of normal go `defer` commands instead of the before/after test hooks used in tools like gomega. This can take getting used to, but it is also the same as normal go code, so shouldn't be time lost. There are some cases where multiple tear-down steps are needed and those steps are more visible now. That's not necessarily bad because these typically are also the most complex tests and have a lot going on. Seeing the various setup requirements shows problem areas that help knowing when debugging.

Signed-off-by: Adrian Cole <adrian@tetrate.io>