Propagates runtime.GOOS as docker env GETENVOY_GOOS to speed up rust (#158)
The current way we build extensions is to mount a volume containing a directory created by "getenvoy extension init" in a "docker run" command. This implies a "bind mount" which is slow on macOS and particularly slow in rust (cargo) builds. This topic has locked our project into an old docker script, and even then it is quite slow.

This change tries to work around the problem by sharing information about the docker host (eg the machine calling getenvoy). What it does is propagates runtime.GOOS as docker env GETENVOY_GOOS. The rust builder image looks at this and optimizes if it is macOS.

Signed-off-by: Adrian Cole <adrian@tetrate.io>