Removes Istio bootstrap
The maintenance and side-effects of Istio bootstrap integration have
proven to be the bulk of effort in maintaining this project, and the
bulk of its size. This removes the feature to free us to focus on Envoy,
which better aligns with the primary charter of the project.

Before:
```bash
$ du -k build/bin/*/amd64/getenvoy
77924	build/bin/darwin/amd64/getenvoy
66692	build/bin/linux/amd64/getenvoy
```

Now:
```bash
$ du -k build/bin/*/amd64/getenvoy
20768	build/bin/darwin/amd64/getenvoy
18428	build/bin/linux/amd64/getenvoy
```

Signed-off-by: Adrian Cole <adrian@tetrate.io>