Cleans signal handling and reduces complexity of "getenvoy run"
Before, we had complex signal handling in both "getenvoy run" and also
normal command invocation. This solves the first part by reducing the
amount of gears used to run the process, and also moving to
NotifyContext for signal interception.

See #172 for more information on NotifyContext, though this change
doesn't remove the osutil functions, yet. That will be a different PR
as it is a different branch of code.

Signed-off-by: Adrian Cole <adrian@tetrate.io>