Removes chrono from rust extension examples
This removes chrono from the logging examples. Using this library
contributes to compilation time and is also unrealistic as logging
libraries typically perform their own timestamps.

This also removes unused build aliases and renames .cargo/config to
.cargo/config.toml per recommendations:
https://doc.rust-lang.org/cargo/reference/config.html

Finally, this removes dead code, as the envoy clock was initialized,
but never used in the rust examples.

Fixes #144

Signed-off-by: Adrian Cole <adrian@tetrate.io>