Removes `getenvoy extension run --envoy-path` and cleans tech debt
`--envoy-path` was mainly used for tests and is removed, but the bulk
of this change is tech debt cleanup.

It is understood that a large amount of tech debt cleanup is here. The
bulk of ideas here were reviewed and @mathetake accepted review burden
before and during this work. We agreed that someone has to clean this
up and it doesn't get easier later.

We introduced options to allow overriding the location of the envoy binary, effectively subverting the idea that getenvoy results in a known working binary. This feature was earlier removed from `getenvoy run` where we formerly sneakily allowed the envoy version to be file path! This completes the job by also removing the `--envoy-path` from `getenvoy extension run`.

As described below, we have other ways to control unit tests now, and we shouldn't bleed that into the CLI arguments unless end-users that don't work here request it.

Below describes tech debt cleanup, things internally that don't effect users directly

Before we had a myriad of different directories and names for similar directories.

* `DebugStore` -> `WorkingDir`
* `TmplDir` -> `WorkingDir`
* `DebugDir` -> `WorkingDir/..` (until we top level the concept and it should be archive dir!)
* `RootDir` -> `HomeDir`
* `WorkspaceDir` -> `ExtensionDir`
* `GetEnvoyBinaryPath` -> `GetEnvoyPath`
* `/source` volume mount in docker -> `ExtensionDir`
* `OutputDir` -> `ExtensionDir`
* `NoPrompt` -> `NoWizard`
* `StylesEnabled` -> `NoColors`

We had several shell scripts that did similar things as well as several test fixtures that also did similar things. Once something repeats a few times, it is now time to abstract!

* `RequireCaptureScript` consolidates the following into :::, using more portable shell syntax.
  * `pkg/binary/envoy/testdata/sleep.sh`
  * `pkg/extension/workspace/toolchain/builtin/testdata/toolchain/docker`
  * `pkg/util/exec/testdata/test_stdio.sh`
  * `pkg/util/exec/testdata/test_trap_sigterm.sh`

As a result, there are few shell scripts and can be manually formatted on change. This also removes the need for `shfmt` from `.bingo` and `Makefile`.

We had many occurrences of interfaces directly coupled to implementation. This assumes a lot including what another implementation might feel like, but more importantly, it creates navigation problems and also type casting everywhere (ex `r.(*Runtime)`. Moreover, it always hurts test coverage. We should stop abstracting until there are 3 implementations of something.

* `binary.Runner` had only one implementation (Runtime) so `pkg/extension/workspace/example/runtime` is gone
* `printer interface` -> `io.Writer`
* `dockerutil.ImageName` -> `fmt.Sprintf`
* `names.Add` -> `append`
* `GetCurrentUser` -> `user.Current`
* `ErrorHandlers` -> explicit code that does the same
* `RunContext` gone
*  `runtime.go` (`RuntimeOption` etc) gone as no longer needed
* `FetchRunner` is not needed to be coupled to `Runtime` and gone
* `ComposableRunE` etc gone as not needed

Running envoy in unit tests makes things like coverage tools difficult. It makes it near impossible to run offline, slows tests, and complicates CI

* no need for the large amount of "status" code used only by debug tests
* no more fetching envoy in GH action for the unit tests
* less checked in envoy config in testate
* "debug" unit tests run significantly faster than before
* "debug" tests can be improved in the future covering far more corner cases than easy to orchestrate with a real envoy process.

Before, we spammed the user with inconsequential knowledge, such as the PID that is already dead. Each line we choose to mark info or higher should be impactful. Even though we override stout/err in most unit tests, our command code defaulted to real stdout/stderr. This is worse than nil for unit tests especially as people shouldn't be bothered with lines of output unless they matter.

* logging of pid when it already died
* redundant logging that includes the function name that's already logged
* Run doesn't default unset to os.Stdout and os.Stderr anymore

Formerly, we had a number of implicit lookup code, which made navigation even more difficult and harder to control. This also hurt test coverage. While it can be tedious to retrofit code to be configurable, mutable global variables are worse. The extra setup/teardown to capture and revert values id distracting. It also doesn't play well in parallel tests.

* `home.go` -> `GlobalOpts.HomeDir`
* `manifest.SetURL()` -> `GlobalOpts.ManifestURL`
* `GetCurrentWorkspace()` -> `GetWorkspaceAt(o.ExtensionDir)`
* `FindWorkspaceDir` is no longer needed
* `OverrideGetCurrentUser` -> `builtin.DefaultDockerUser`
* `globals.NoPrompt` -> `GlobalOpts.NoWizard`
* `globals.NoColors` -> `GlobalOpts.NoColors`
* PWD overrides for extension commands -> `GlobalOpts.ExtensionDir`
* PATH overrides for docker -> `GlobalOpts.DockerPath`
* `envoy.tmpl.yaml` -> `envoy.yaml`