Removes `getenvoy extension run --envoy-path` and cleans tech debt (#195)
`--envoy-path` was mainly used for tests and is removed, but the bulk
of this change is tech debt cleanup.

We introduced options to allow overriding the location of the envoy binary, effectively subverting the idea that getenvoy results in a known working binary. This feature was earlier removed from `getenvoy run` where we formerly sneakily allowed the envoy version to be file path! This completes the job by also removing the `--envoy-path` from `getenvoy extension run`.

Signed-off-by: Adrian Cole <adrian@tetrate.io>
Co-authored-by: Takeshi Yoneda <takeshi@tetrate.io>