Removes all extension code (#200)
A myriad of maintenance issues and conversations led to the summary
outcome that the "extension" subsystem is not a right fit. Rather than
continue weeks of refactoring and otherwise quality corrections, this
removes all extension (WASM) features.

The time otherwise spent on the >15k lines of code and configuration
can go towards improving getenvoy. Improving getenvoy will be far
easier when it isn't subject to hour turn-arounds on rust builds or
the complexity of embedding Docker.

Instead, we will suggest developers use the underlying SDKs and their
corresponding developer workflows. Over time, we may tighten integration
here with WASM in less coupled ways, for example integrating existing
E2E tests from TinyGo is far simpler than embedding builds, and easier
on CI as well.

Meanwhile, the supported way to run any extension (Lua or WASM), is to
pass a configuration file to the `getenvoy run` command.

Signed-off-by: Adrian Cole <adrian@tetrate.io>
Co-authored-by: Takeshi Yoneda <takeshi@tetrate.io>