Updates to latest go and deps, restores "indirect" (#206)
This updates versions to latest with no visible impact, while also
cleaning up the workflow files wrt centralizing the go version.

Before, we accidentally didn't have indirect dependencies in our go.mod
file. This hid why our binary is so large (>12M). This restores the
inconvenient information so that it becomes easier to remove the bulk.

Signed-off-by: Adrian Cole <adrian@tetrate.io>