Removes docker image for getenvoy
We added `getenvoy/getenvoy` when testing istio. This was curiously
composed including pinning to Envoy 1.12. Real use of `getenvoy` would
be odd to run Envoy in Docker because there's already an envoy image.

This removes this as if it were used in practice, someone would have
complained about it being targetted to a version from 2019.

Signed-off-by: Adrian Cole <adrian@tetrate.io>