Stops admin fetching on first error
Before, we added a dependency on multi-error to allow multiple problems
to accumulate before before returning control to the user. There's
little reason to allow partial failure across admin URLS, especially
when a partial failure still quits with an error status.

This removes the dependency and works normally instead. This also fixes
syntax where formerly we used defer in a for loop.

Signed-off-by: Adrian Cole <adrian@tetrate.io>