Refactors logging to not bother users on expected errors (#212)
Before, we had debug logs that were not possible to enable, as well logs
that bothered users with situations they could not action (such as vague
details about child process signals.). This stops logging on aspects the
user cannot control. It tightens logging so that if there's something we
don't expect, the user will see it.

For example, we know certain OS utilities are unimplemented, so we skip
writing files or fields when that happens. However, we don't bother
users. If there is something unexpected, it is logged at most once per
debug utility. We do not log anything anymore about a possibility of
someone doing "kill -9" and leaving the child envoy around. This is pure
distraction for most people.