Removes incomplete and untested lsof embedding
Issue #47 requested and #68 partially implemented lsof on shutdown. This
has had some bad effects on the codebase. Particularly it is the most
complicated part of it. Due to it hanging, we had to wrap it in a
timeout block. Also, it is unusuable on darwin being unimplemented. No
testing was done, so only visually looking at code can you tell that
fields are missing. It also uses platform-specific code to get inum
info, which means it won't compile on windows. In summary:

* partially might work on ubuntu
* doesn't work on darwin
* doesn't compile on windows
* completely untested

While the original goal was no dependencies, the library used doesn't
support lsof enough well enough to call this a feature. It is far
simpler and more supportable to invoke lsof as at least that's available
on mac. In any case here or via lsof bin, doing file collection like
this is a relatively slow and unsafe task. So, this removes "lsof"
pre-terminate until some end users ask for it, and unless gopsutil
becomes better at it, likely invoking the binary would be the way out.

Signed-off-by: Adrian Cole <adrian@tetrate.io>