Refactors JSON schema and adds SHA256 verification of tarballs (#281)
This change allows us increased confidence that tarballs downloaded were
what we expected. Notably, `getenvoy use N.N.N` will fail now, if the
`sha256sum` doesn't match what's in the json key for that filename.

This refactors out Envoy-specific terminology from the JSON schema used
to validate envoy-versions.json. It is now called
release-versions-schema.json.

This adds a new key sha256sums in the schema and backfills data for the
existing versions.

The key name "sha256sums" comes from the `sha256sum` command. This is
SHA-256 instead of SHA-512 as it is a ubuquitous hash algo, for example,
used by default also in goreleaser.

This doesn't use a more generic term like "hash" because doing so would
cause unnecessary confusion on what hash it is (ex MD-5?).

The map "sha256sums" allows for easy indexing, and using the filename
instead of the URL allows less changes when rehosting the exact same
files.

Signed-off-by: Adrian Cole <adrian@tetrate.io>
Co-authored-by: Takeshi Yoneda <takeshi@tetrate.io>