Adds Windows installer
This adds a Windows installer (msi) using the subset of WIX features
supported by https://wiki.gnome.org/msitools.

Dodging unsupported features brings DevX because we can create and test
the tool on OS/x.

Currently, this only verifies that the installer built on all operating
systems. This is an attempt to keep the amount of change to a minimum.

Signed-off-by: Adrian Cole <adrian@tetrate.io>