Uses make in windows (#325)
This consolidates the build by using the same matrix for Linux, MacOS
and Windows. To do the latter requires installation of CygWin, and to
get GHA to use that requires overriding the `shell`. One other quirk is
that when you are running tasks via cygwin, it resets the CWD. To avoid
that, this adds the GITHUB_WORKSPACE to `.bashrc`. Finally, this dodges
various newline related problems and comments why.

Unrelated changes including updating syntax that allows us to
break-apart actions/cache and fixing some lint errors that the build
never caught because lint wasn't run on Windows.

Signed-off-by: Adrian Cole <adrian@tetrate.io>