test: Normalize file permission when comparing files
This patch put mask on file mode to ignore the group write policy bit of
the compared files. This is required since for example on Linux when
running as a normal user, the umask default value is 002 affecting the
file permission changes for group write policy of files checked by "git
clone" from a remote git repository.

Signed-off-by: Dhi Aurrahman <dio@tetrate.io>