Add FuncEVersions interface
This patch introduces FuncEVersions as an interface for easier testing,
especially when simulating failure on getting the remote Envoy release
versions map.

This also provides the default implementation of FuncEVersions, which
converts the existing functionality via Get() method and also a way to
find a latest patch given Envoy release versions map and a minor
version.

Signed-off-by: Dhi Aurrahman <dio@rockybars.com>