Add tests calling signal to func-e
This is a follow-up for
https://github.com/tetratelabs/func-e/pull/371#pullrequestreview-752322738
to test all possible signals to func-e.

A minor change on checking the stderr: this uses scanner instead of peek
on the buffer since for some reason redirecting cmd.Stderr to a buffer
doesn't give complete stderr after the process is terminated. Hence here
we use scanner to io.ReadCloser of cmd.StderrPipe() instead.

Signed-off-by: Dhi Aurrahman <dio@rockybars.com>