Removes go version maintenance (#379)
This removes the hard requirement to build with exactly the same version
of go as indicated in the go.mod file. This allows more flexibility when
new versions of go (such as 1.18) are released.

This also removes the requirement to track down to the patch of go.
Instead, this builds with the latest patch of go that's available at the
time.

Go versions aren't often released, but this removes maintenance each
time that happens. It also allows us to more easily cache golang
installs in our Docker image, and without tight coupling to what GitHub
actions uses.

Signed-off-by: Adrian Cole <adrian@tetrate.io>