Matches minor revision of go of CentOS image with GHA runner
This matches the minor revision of go with the current GitHub Actions
runner instead of looking up latest using gimme. Doing so removes
complexity around what gimme is, and why figuring out the latest go
revision for a minor isn't straight-forward [1].

[1] Unlike a lot of projects, go does not attach binaries to releases.
This makes a potential race between a tag being pushed and binaries
being available. Moreover, there are release tags for release
canditates, so filtering these out requires logic. The way to get the
best known latest revision is to parse https://golang.org/dl, as that
is the most official source of what is the latest minor. However,
parsing that is tricky, so if we did, we might as well re-use gimme.

Ex.
```yaml
        run: |
          GIMME_SCRIPT=https://raw.githubusercontent.com/travis-ci/gimme/master/gimme
          GO_MINOR_REVISION=$(curl -sSL ${GIMME_SCRIPT} | GIMME_GO_VERSION=${{ env.GO_VERSION }}.x bash -s -- -r)
```

The above is for reference only because after this change, we don't
lookup a version of go anymore.

Signed-off-by: Adrian Cole <adrian@tetrate.io>