Ensures Ubuntu runs the same version of go as CentOS (#385)
This avoids any version differences between running e2e tests via Docker
(ex for CentOS) vs a normal run (Ubuntu, MacOS, etc). It accomplishes
this by using the built-in GOROOT variable.

go.mod controls which GOROOT is used, and in doing so, this removes
maintenance.

This includes a temporary workaround because MacOS does not have the
same variable as Windows and Ubuntu. Once the below is solved, we can
simply delete the workaround.