Treats versions constants consistently as strings to reduce conversions (#400)
Before, we tried to force embedded versions as version.Version type.
This became less useful when we introduced support for minor versions
(which don't satisfy the same format). This goes back to strings as it
is now a cleaner option. The type is only used to relate the JSON
structure together, and since that still helps, it is left in.

Signed-off-by: Adrian Cole <adrian@tetrate.io>