Adds union type to handle the two valid version arguments (#402)
Recently, we added the ability to request the latest patch, by leaving
out the patch component of the version. Ex "1.18" instead of "1.18.3".
The implemenation of this resulted in inconsistency and duplication.

This adds a union type `version.Version` which is an interface to cover
both use cases. This also cleans up some logic and regex duplication.

Signed-off-by: Adrian Cole <adrian@tetrate.io>