Cleans up version selection logic (#403)
This removes an api for acknowledgement of the initial state. Doing so
also reduces file IO when environment variables select the version, as
in this case no default version file will be created.

This also fixes a bug where we had a false-positive on version selection.
Versions are platform-specific so we need to only be ok when the
current platform's tarball is available.

Next, this alerts the user when a remote lookup is happening. This was
missed before as we didn't let users know we are remotely resolving the
latest version.

Finally, this ensures there's enough information in warnings and errors
to troubleshoot a missing release. Particularly, there are cases where a
version is available, but not across the board. For example, linux exists,
but darwin does not.

Signed-off-by: Adrian Cole <adrian@tetrate.io>