Ensures last_known_envoy.txt doesn't reference an inconsistent version
This changes the logic that verfies last_known_envoy.txt such that it
only expects updates when all supported platforms are available.

Before, this would update when the first platform was available, which
would break development (ex linux out before macOS).

Note: this doesn't address the similar problem at runtime, and a variant
of the problem in the e2e tests. This PR is standalone to highlight the
switch from using `jq` to go for linting.

See #393

Signed-off-by: Adrian Cole <adrian@tetrate.io>