Fixes unnaturally high test coverage by excluding test utilities
Before, our test coverage was accidentally 0.5% higher than it should be
due to inclusion of packages not in the binary (notably test utilities).

This fixes it in native make, which helps us maintain what is main or
not in one place. We use make-derived source lists in order to taint
targets properly.

Signed-off-by: Adrian Cole <adrian@tetrate.io>