Fixes patterns used to weed out test sources
Previously, we used an invalid pattern style as regex isn't supported.
It was interpreted as a character class and resulted in invalid
filtering. Specifically, this prevented the "moreos" package from
showing up in coverage results. To fix this portably, we enumerate known
depths of what we don't want. Doing this looks explicit, but it
shouldn't result in high maintenance as this project isn't likely to
have deep package hierarchies.

Signed-off-by: Adrian Cole <adrian@tetrate.io>