Updates to use GitHub Action for codecov
The codecov GitHub Action replaces the uploader script for projects that
are either mostly amd64 or can consolidate uploads into one stage that
runs on amd64. func-e has no significant arm64 source files, so can use
this action instead of the uploader script.

Thanks to @anuraaga for noticing the uploader script is deprecated.

Signed-off-by: Adrian Cole <adrian@tetrate.io>