packaging: use native WiX toolchain on Windows
Cross-platform msitools do not support the `Environment` directive,
which should be the proper way to set up the PATH environment variable,
so we adapt the Makefile to assume a Windows environment when building.

`make dist` now runs on all platforms, but on Windows, assuming
`candle.exe`, `light.exe` and GNU `tar` are installed, it builds
all tarballs, zip and msi packages.

Notice this was already supported, but it still relied on `SETX`
to support building cross-platform.

On other platforms it will only build the tarballs.

Signed-off-by: Edoardo Vacchi <evacchi@users.noreply.github.com>